+++ src/Entity/Block.php
@@ -13,63 +13,18 @@ declare(strict_types=1);
 
 namespace MonsieurBiz\SyliusCmsPagePlugin\Entity;
 
-use DateTimeInterface;
 use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Sylius\Component\Channel\Model\ChannelInterface;
-use Sylius\Component\Resource\Model\TimestampableTrait;
-use Sylius\Component\Resource\Model\ToggleableTrait;
 use Sylius\Component\Resource\Model\TranslatableTrait;
 use Sylius\Component\Resource\Model\TranslationInterface;
 use Webmozart\Assert\Assert;
 
-class Block implements BlockInterface
+class Block extends ContentManagement implements BlockInterface
 {
-    use TimestampableTrait;
-    use ToggleableTrait;
     use TranslatableTrait {
         __construct as private initializeTranslationsCollection;
         getTranslation as private doGetTranslation;
     }
 
-    /**
-     * @var int|null
-     */
-    protected $id;
-
-    /**
-     * @var bool
-     */
-    protected $enabled = true;
-
-    /**
-     * @var string|null
-     */
-    protected $code;
-
-    /**
-     * @var Collection<int, ChannelInterface>
-     */
-    protected $channels;
-
-    /**
-     * @var DateTimeInterface|null
-     */
-    protected $createdAt;
-
-    /**
-     * @var DateTimeInterface|null
-     */
-    protected $updatedAt;
-
-    /**
-     * @var string
-     */
-    protected $title;
-
-    /**
-     * Block constructor.
-     */
     public function __construct()
     {
         $this->initializeTranslationsCollection();
@@ -77,85 +32,9 @@ class Block implements BlockInterface
     }
 
     /**
-     * @return int|null
+     * @var string
      */
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return Collection<int, ChannelInterface>
-     */
-    public function getChannels(): Collection
-    {
-        return $this->channels;
-    }
-
-    /**
-     * @param ChannelInterface $channel
-     */
-    public function addChannel(ChannelInterface $channel): void
-    {
-        $this->channels->add($channel);
-    }
-
-    /**
-     * @param ChannelInterface $channel
-     */
-    public function removeChannel(ChannelInterface $channel): void
-    {
-        $this->channels->removeElement($channel);
-    }
-
-    public function initializeChannelsCollection(): void
-    {
-        $this->channels = new ArrayCollection();
-    }
-
-    public function hasChannel(ChannelInterface $channel): bool
-    {
-        return $this->channels->contains($channel);
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getContent(): ?string
-    {
-        return $this->getTranslation()->getContent();
-    }
-
-    /**
-     * @param string|null $content
-     *
-     * @return void
-     */
-    public function setContent(?string $content): void
-    {
-        $this->getTranslation()->setContent($content);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function createTranslation(): BlockTranslation
-    {
-        return new BlockTranslation();
-    }
-
-    /**
-     * @param string|null $locale
-     *
-     * @return BlockTranslationInterface
-     */
-    public function getTranslation(?string $locale = null): TranslationInterface
-    {
-        $translation = $this->doGetTranslation($locale);
-        Assert::isInstanceOf($translation, BlockTranslationInterface::class);
-
-        return $translation;
-    }
+    protected $title;
 
     /**
      * @return string
@@ -174,18 +53,29 @@ class Block implements BlockInterface
     }
 
     /**
-     * @return string|null
+     * {@inheritdoc}
      */
-    public function getCode(): ?string
+    public function createTranslation(): BlockTranslation
     {
-        return $this->code;
+        return new BlockTranslation();
     }
 
     /**
-     * @param string|null $code
+     * @param string|null $locale
+     *
+     * @return BlockTranslationInterface
      */
-    public function setCode(?string $code): void
+    public function getTranslation(?string $locale = null): TranslationInterface
     {
-        $this->code = $code;
+        $translation = $this->doGetTranslation($locale);
+        Assert::isInstanceOf($translation, BlockTranslationInterface::class);
+
+        return $translation;
     }
+
+    public function initializeChannelsCollection(): void
+    {
+        $this->channels = new ArrayCollection();
+    }
+
 }
