+++ src/Entity/Page.php
@@ -13,120 +13,22 @@ declare(strict_types=1);
 
 namespace MonsieurBiz\SyliusCmsPagePlugin\Entity;
 
-use DateTimeInterface;
 use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Sylius\Component\Channel\Model\ChannelInterface;
-use Sylius\Component\Resource\Model\TimestampableTrait;
-use Sylius\Component\Resource\Model\ToggleableTrait;
 use Sylius\Component\Resource\Model\TranslatableTrait;
 use Sylius\Component\Resource\Model\TranslationInterface;
 use Webmozart\Assert\Assert;
 
-class Page implements PageInterface
+class Page extends ContentManagement implements PageInterface
 {
-    use TimestampableTrait;
-    use ToggleableTrait;
     use TranslatableTrait {
         __construct as private initializeTranslationsCollection;
         getTranslation as private doGetTranslation;
     }
 
-    /**
-     * @var int|null
-     */
-    protected $id;
-
-    /**
-     * @var bool
-     */
-    protected $enabled = true;
-
-    /**
-     * @var string|null
-     */
-    protected $code;
-
-    /**
-     * @var Collection<int, ChannelInterface>
-     */
-    protected $channels;
-
-    /**
-     * @var DateTimeInterface|null
-     */
-    protected $createdAt;
-
-    /**
-     * @var DateTimeInterface|null
-     */
-    protected $updatedAt;
-
-    /**
-     * Page constructor.
-     */
     public function __construct()
     {
         $this->initializeTranslationsCollection();
         $this->initializeChannelsCollection();
-    }
-
-    /**
-     * @return int|null
-     */
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getCode(): ?string
-    {
-        return $this->code;
-    }
-
-    /**
-     * @param string|null $code
-     */
-    public function setCode(?string $code): void
-    {
-        $this->code = $code;
-    }
-
-    /**
-     * @return Collection<int, ChannelInterface>
-     */
-    public function getChannels(): Collection
-    {
-        return $this->channels;
-    }
-
-    /**
-     * @param ChannelInterface $channel
-     */
-    public function addChannel(ChannelInterface $channel): void
-    {
-        $this->channels->add($channel);
-    }
-
-    /**
-     * @param ChannelInterface $channel
-     */
-    public function removeChannel(ChannelInterface $channel): void
-    {
-        $this->channels->removeElement($channel);
-    }
-
-    public function initializeChannelsCollection(): void
-    {
-        $this->channels = new ArrayCollection();
-    }
-
-    public function hasChannel(ChannelInterface $channel): bool
-    {
-        return $this->channels->contains($channel);
     }
 
     /**
@@ -145,24 +47,6 @@ class Page implements PageInterface
     public function setTitle(?string $title): void
     {
         $this->getTranslation()->setTitle($title);
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getContent(): ?string
-    {
-        return $this->getTranslation()->getContent();
-    }
-
-    /**
-     * @param string|null $content
-     *
-     * @return void
-     */
-    public function setContent(?string $content): void
-    {
-        $this->getTranslation()->setContent($content);
     }
 
     /**
@@ -255,4 +139,10 @@ class Page implements PageInterface
 
         return $translation;
     }
+
+    public function initializeChannelsCollection(): void
+    {
+        $this->channels = new ArrayCollection();
+    }
+
 }
